move helper to a parent class
